requirements.txt
# Sinew — Person B dependencies
# Install with: pip install -r requirements.txt

# Core
flask>=3.0
requests>=2.31
pyyaml>=6.0
python-dotenv>=1.0

# AI
anthropic>=0.40

# Vision
opencv-python>=4.9
numpy>=1.26
Pillow>=10.0

# Voice
faster-whisper>=1.0
pyaudio>=0.2.14

# TTS
pyttsx3>=2.90

# Calibration GUI (Person A, but useful for B to install to avoid import errors)
# PyQt5>=5.15
